Adding live data provider in nod module
- Parses raw buffer format
- Converts to xviz data
- Sends xviz data to visualizer

+++ modules/parser/src/parsers/parse-xviz-stream.js
@@ -516,7 +516,7 @@ function joinObjectPointCloudsToTypedArrays(objects) {
 
   let vertexColorStride = null;
 
-  const positions = new Float32Array(numInstances * 3);
+  const positions = new Float64Array(numInstances * 3);
   const colors = new Uint8ClampedArray(numInstances * 4);
 
   // Store object ids to enable recoloring.
